Add ResponseComparison model for parity check
As part of the API parity check we will need to restore the results for each
lead provider/path that we are calling.

Add `ResponseComparison` model to store the comparison results. If the response
is equal we will clear the response bodies prior to saving to avoid storing
large amounts of data in the database.