Add ParityCheck service
The service will perform API requests against the ECF and NPQ services, storing
the results in a `ResponseComparison`. We use the `TokenProvider` service to
generate tokens at the start of the check and we benchmark the reqest durations
to get an idea of performance between the two services.

Currently we are only testing the statements endpoint.

Comparing/diffing the results will be left to a future PR that adds the UI.

There are no differences reported in the statement endpoints after running in
migration.