Add token generation service for parity check (#1810)
* Add ECF models for ApiToken generation

Bring across the models from ECF that we need to be able to generate API tokens
for ECF in NPQ registration.

I haven't improved the test coverage here, just a quick lift and shift from ECF.

* Add token generation service for ECF/NPQ

As part of the parity check we will need to generate API tokens for lead
providers in ECF and NPQ reg on the fly to be able to make requests to the API
as these providers.

Add service to generate tokens for each lead provider.