Add basic parity check for API endpoints (#1817)
* Add ResponseComparison model for parity check

As part of the API parity check we will need to restore the results for each
lead provider/path that we are calling.

Add `ResponseComparison` model to store the comparison results. If the response
is equal we will clear the response bodies prior to saving to avoid storing
large amounts of data in the database.

* Add ParityCheck service

The service will perform API requests against the ECF and NPQ services, storing
the results in a `ResponseComparison`. We use the `TokenProvider` service to
generate tokens at the start of the check and we benchmark the reqest durations
to get an idea of performance between the two services.

Currently we are only testing the statements endpoint.

Comparing/diffing the results will be left to a future PR that adds the UI.

There are no differences reported in the statement endpoints after running in
migration.

* Preload lead provider tokens in parity check service

If we don't do this, the first request will also load all the toakens which
takes a few seconds and skews the request benchmarking.