Add completion and details to parity check UI
Fleshes out the bare-bones parity check UI with an overview list of the paths
called for each provider on the completion page. If the request
responses/status codes are different you can click on the path to view a
detailed report of the comparison along with a diff of the response bodies.
Also includes a basic response time comparison chart that plots the averages
and min/max for each path.

Add detailed repsonse comparison action/view.

Add `diffy` to compare response bodies.

Update CSRP to allow chart.js from a CDN (as its only used on this page and it
will be deleted shortly I figured this was the best way to include it/ensure
its removed).

Remove `equal?` from `ResponseComparison` as this is a system method it started
giving issues; we can just negate `different?` in its place.