Update parity check UI to support pagination
Display comparison results grouped by description on the completion page (so
comparisons with the same method/path and different page will be grouped
together).

On the completion page we show one entry for paginated comparisons (if there
is one difference across any page then its marked as different).

On the detail page we only show the different pages in the result, along
with a summary of the average performance/response times.

The per page value is intentionally set as a low value for now so that we can see the pagination work on statements.