Add parity check UI (#1829)
* Add basic UI for parity checks

Add a basic UI to kick off a parity check and view the progress/completion
status.

Add `ParityCheckJob` to run parity checks in the background.

Add methods to `ParityCheck` service for tracking progress.

Add `ParityChecksController` and associated views; no details of the check
results are surfaced yet, just the functionality to start and view the overall
status.

DRY up parity check tests with `migration_helper`.

* Add completion and details to parity check UI

Fleshes out the bare-bones parity check UI with an overview list of the paths
called for each provider on the completion page. If the request
responses/status codes are different you can click on the path to view a
detailed report of the comparison along with a diff of the response bodies.
Also includes a basic response time comparison chart that plots the averages
and min/max for each path.

Add detailed repsonse comparison action/view.

Add `diffy` to compare response bodies.

Update CSRP to allow chart.js from a CDN (as its only used on this page and it
will be deleted shortly I figured this was the best way to include it/ensure
its removed).

Remove `equal?` from `ResponseComparison` as this is a system method it started
giving issues; we can just negate `different?` in its place.

* Consolidate `npq_separation` request specs

We have these under both `spec/controllers` and `spec/requests`; we agreed to
use `spec/requests` going forward. We should also look to clean up the file
naming (to remove `_controller`) but as there are a lot in this format it would
be best done in another PR.