Support pagination and filtering in parity check service (#1840)
* Add page attribute to ResponseComparison

We want to support pagination when calling certain endpoints as party of the
parity check. To make it more managable, we're going to store the results of
each page as a separate `ResponseComparison` instance.

Add `page` attribute to `ResponseComparison` model.

* Extract Client from ParityCheck service

To avoid overcomplicating the `ParityCheck` service, the request/response logic
has been extracted to a separate `Client` model. This model will be expanded to
support pagination and any other request/response behaviour.

* Support pagination

Update the `Client` to support paginating if the `options[:paginate]` us
specified.

It will paginate until there are no pages left for either NPQ or ECF
(determined by the resulting `data` count being less than the `per_page`).

* Update parity check UI to support pagination

Display comparison results grouped by description on the completion page (so
comparisons with the same method/path and different page will be grouped
together).

On the completion page we show one entry for paginated comparisons (if there
is one difference across any page then its marked as different).

On the detail page we only show the different pages in the result, along
with a summary of the average performance/response times.

The per page value is intentionally set as a low value for now so that we can see the pagination work on statements.