Add POST endpoints to parity check
Support `post` requests in the parity check tool. In order to do this we expand
the `Client` to support POST requests and allow specifying a `payload` in the
YAML for a path. This can either be a hash, which will be sent as the request
body or it can be a method name that can construct a dynamic payload.