Treat non-200 responses as needs review
Currently, if both services return a non-200 response and its the same then it
will be flagged as equal. Whilst this is technically correct its not very
helpful as it can hide issues with the calls themselves.

Instead, we want to raise these as 'unexpected' results so that we get an
indiciation something is wrong.