Add POST/PUT endpoint support to parity check (#1874)
* Add POST endpoints to parity check

Support `post` requests in the parity check tool. In order to do this we expand
the `Client` to support POST requests and allow specifying a `payload` in the
YAML for a path. This can either be a hash, which will be sent as the request
body or it can be a method name that can construct a dynamic payload.

* Add PUT endpoints to parity check

Support `put` requests in the parity check tool. In order to do this we expand
the `Client` to support PUT requests and add specific use cases to the YAML
file/client.

* Memoize request body

To avoid issues with dynamic payloads and ensure both the NPQ and ECF requests
use the exact same body we can memoize it.

* Exclude fields that will be different in POST/PUT

Mainly the updated/created attributes that will have different values.

Minor bug fixes to some of the `id` queries off the back of errors we got
during testing.

* Fix feature specs

* Treat non-200 responses as needs review

Currently, if both services return a non-200 response and its the same then it
will be flagged as equal. Whilst this is technically correct its not very
helpful as it can hide issues with the calls themselves.

Instead, we want to raise these as 'unexpected' results so that we get an
indiciation something is wrong.

* Minor bug fixes and tweaks

Small changes to get the POST/PUT comparison results looking better/more reliable.