Set swagger URL to current environment
We currently hardcode the swagger URL; it is packaged into the Docker image and
we deploy the same image to all environments.

We want to have the current domain in the URL for whichever environment the
application is deployed to. In order to do this, we can replace the URL on
application boot. It's not ideal that we do this on every boot, but the
alternative would involve a separate docker image for each environment which
would be a lot harder to manage.