Split parity check across multiple workers
As the workers only have 4 CPU cores we found multithreading to not be that
efficient.

Instead, this commit will split the requests across multiple workers. It uses a
mutex on the database to ensure the last worker to finish will finalise the
parity check.

The testing becomes quite complicated/contrived due to relying on accessing
`Delayed::Job` directly for the check status (so we can't use the `:test` queue
adapter).

I'm not sure why but I couldn't override the default `test` queue adapter in
the feature specs to use `delayed_job` in order to test it end-to-end (it
always ends up using the `test` queue adapter for some reason). I had to ditch
some tests here as they can no longer pass (completion relies on querying the
state of `Delayed::Job` directly now).