Set swagger URL to current environment (#1891)
* Set swagger URL to current environment

We currently hardcode the swagger URL; it is packaged into the Docker image and
we deploy the same image to all environments.

We want to have the current domain in the URL for whichever environment the
application is deployed to. In order to do this, we can replace the URL on
application boot. It's not ideal that we do this on every boot, but the
alternative would involve a separate docker image for each environment which
would be a lot harder to manage.

* Fix review app domain env var

Fixes a typo in the NPQ reg domain name.

* Pass external domain into environment

Pass the external domain into the environment so we can use it in the swagger
docs. Unlike ECF we don't seem to have external domains for non-prod
environments so I've just used the internal domains for now.