Ensure disabled providers are included in serializers
We have a default scope on the providers that omits disabled providers by
default. This is causing the providers to be `nil` in the serializers, but
the tests are passing as the scope is bypassed due to the providers being
preloaded in the spec already.

Update specs to ensure the application is reloaded and the default scopes
are correctly applied.

Update `Application` to descope via new `belongs_to` associations suffixed
with `including_disabled` so we return disabled providers when accessed via
the application.

Improve test coverage of nullable values in `ApplicationCsvSerializer`.