Update `updated_at` of application if provider only in NPQ reg
If we migrate an application from ECF that doesn't have a private childcare
provider and/or itt provider, but the corresponding application in NPQ reg
does have one of these set, we want to ensure the `updated_at` is updated to
the current time so that the change will be surfaced to providers.

I've opted to include disabled providers here for consistency, but all the
current providers in NPQ reg should be active.