Skip user touch during migration when application changes
When an application changes and the `lead_provider_approval_status` is
updated we are touching the user `updated_at`. We don't want to do this
during the migration as it is causing user timestamps to change and
differences in the parity check.

Add `skip_touch_user_if_changed` attribute to `Application` and set it to
`true` for the migration `update!` call.