Fix bug with providers not populating in serializers (#1912)
* Ensure disabled providers are included in serializers

We have a default scope on the providers that omits disabled providers by
default. This is causing the providers to be `nil` in the serializers, but
the tests are passing as the scope is bypassed due to the providers being
preloaded in the spec already.

Update specs to ensure the application is reloaded and the default scopes
are correctly applied.

Update `Application` to descope via new `belongs_to` associations suffixed
with `including_disabled` so we return disabled providers when accessed via
the application.

Improve test coverage of nullable values in `ApplicationCsvSerializer`.

* Update `updated_at` of application if provider only in NPQ reg

If we migrate an application from ECF that doesn't have a private childcare
provider and/or itt provider, but the corresponding application in NPQ reg
does have one of these set, we want to ensure the `updated_at` is updated to
the current time so that the change will be surfaced to providers.

I've opted to include disabled providers here for consistency, but all the
current providers in NPQ reg should be active.