Bump web apps in prod/migration (#1964)
* Bump web apps to 12

We want to get the performance on par with ECF; during testing we're a bit
slower and it seems the web apps are the bottleneck; going to 12 during
performance testing meant NPQ could scale to meet a similar throughput ECF can
manage.

* Bump migration workers to 30

Bump to 30 to match migration environment ahead of running migration in production.

* Bump workers to reflect web app changes

Move to 6 workers now we are running 12 web apps.

* Scale down migration

We need these resources for production and we've finished performance testing now.