Fix race condition with `run_once`
There is a potential race condition in `run_once` migrator logic whereby if the
`run_once` depends on data modified as part of the migration it might happen
too early.

To avoid this, we can change the behaviour so the single-run code goes at the
end and is ran by the last worker to complete. This way it ensures the
migration for the model has fully completed before executing.