Prevent double-seeding the review apps
We currently run `db:seed` when the container boots and then immediately run
a `db:seed:replant` in the following github action step. This can cause the
healthcheck to fail as the courses are purged until the replant re-adds them.
It results in the review app deploy being flakey.

Use existing rake task to seed courses on container boot (which will be fast
and allow the healthcheck to pass) and then run a replant on the following
step to add the remaining data once the container has started.