Re-instate SHA check on deploy
Re-instate the SHA check on deploy that was removed in case it was causing
flakiness; we were seeing failures on the SHA check, but they appear to be
related to other issues with the pipeline.

Check the correct SHA is deployed, once per second for a maximum of 10 seconds
(to give time for the old pods to drain and new pods to spin up).