Add `significantly_updated_at` attribute to `User`
We have an `updated_at` on the `User` model, however this is updated frequently
and for actions that lead providers don't need to care about (such as when a
user signs in).

We want to add a second `updarted_at` attribute that only updates when
something significant changes with the `User`, such as their name, TRN, etc.

As the `Application` can touch a `User` we also need this attribute to update
when the user is touched.

I opted for `significantly_updated_at` as the attribute name as opposed to
`api_updated_at` to try and keep it more generic in case we want to surface it
elsewhere.

Index the field to ensure endpoints that filter by `updated_at` will be performant when we switch them to `significantly_updated_at`.