Remove migrators and associated code
Once we have ran the migration in production we no longer need/want the
migrators and associated code in the codebase (we will no longer have a need
for it and it becomes a potential liability).

Remove all migrators and associated code.

Leave parity check tool in place; this will be removed in a separate commit.