Fix bug in filtering in `Participants::Query`
When we filter on `includes` with ActiveRecord the included table is filtered
down to the records that match the conditional. For the `Participants::Query`
this means when filtering by `updated_since` the related applications/id
changes are filtered as well as the base users. Instead, we want the users to
be filtered but retain all the applications and id changes in the preloaded
association.

To achieve this we can instead use a subquery to find the applications/id
changes updated within the timeframe and filter on the selected `user_id` of
those.

This will be a little less efficient, but is heaps simpler than trying to do an
additional join on applications/id changes to try and get Rails to retain them
all after filtering.