Add `significantly_updated_at` attribute to `User` (#1991)
* Add `significantly_updated_at` attribute to `User`

We have an `updated_at` on the `User` model, however this is updated frequently
and for actions that lead providers don't need to care about (such as when a
user signs in).

We want to add a second `updarted_at` attribute that only updates when
something significant changes with the `User`, such as their name, TRN, etc.

As the `Application` can touch a `User` we also need this attribute to update
when the user is touched.

I opted for `significantly_updated_at` as the attribute name as opposed to
`api_updated_at` to try and keep it more generic in case we want to surface it
elsewhere.

Index the field to ensure endpoints that filter by `updated_at` will be performant when we switch them to `significantly_updated_at`.

* Switch `User` migrator to populate `significantly_updated_at`

We want to populate `significantly_udpated_at` instead of `updated_at` as we
will be exposing this attribute in the serializers for the lead provider API.

* Update queries to use `significantly_udpated_at`

We want to filter by the `significantly_updated_at` attribute instead of the
`updated_at` attribute when querying users for the lead provider API. This will
prevent insignificant changes to a `User` resulting in them being re-surfaced
in the API. Update `Participants::Query` and `Applications::Query` to use
`significantly_updated_at`.

* Update `ParticipantSerializer` to use `significant_updated_at`

We want to surface the `sigificant_updated_at` attribute of participants in the
lead provider API responses so that providers only see changes/updates when
something they care about has changed with the user.

* Fix failing feature specs

Exclude `significantly_updated_at` from these feature specs.

* Migrate `updated_at` as well as `significantly_updated_at`

We want to retain updating `updated_at` on users as well as populating the new
`significantly_updated_at` during the migration.

* Explicitly skip touching `significantly_updated_at` on migration

We want to ensure we don't touch `significantly_updated_at` from the callback
in `User` when performing a migration.

* Backfill the `significantly_updated_at` from `updated_at`

We want to backfill the `significantly_updated_at` using the `updated_at`
attribute of each user.