Fix `Participants::Query` filtering on `training_status` (#2005)
When filtering participants on `training_status` we were still returning all
`applications` in the resulting relationship of each `User`. To be consistent
with ECF we want to only return the applications that match the
`training_status` we are filtering to.

Add test to verify that filtering by `from_participant_id` behaves in the same
way; we only want participant id changes that match the filter to be in the
results.