Remove migrators and associated code (#1993)
* Remove migrators and associated code

Once we have ran the migration in production we no longer need/want the
migrators and associated code in the codebase (we will no longer have a need
for it and it becomes a potential liability).

Remove all migrators and associated code.

Leave parity check tool in place; this will be removed in a separate commit.

* Remove nightly migration GitHub action

Remove the GitHub action that runs and reports on a nightly migration.

* Drop DataMigration model/table

* Remove migration workers

We no longer need these workers as we're finished with the migration.