Fix rate limiting (#2022)
* Fix rate limiting

We had a catch-all/generic rate limit that was set to 300 requests/5 minutes.
This is lower than all of our other throttles in RackAttack and was causing API
requests to be throttled incorrectly.

Instead, we want to add an additional throttle for non-API requests at 300
requests/minute and a catch all/backstop throttle that will ensure no request
to the service will exceed 1500 requests/5 minutes.

* Explicitly rate limit public API paths

These would fallback to the catchall previously, but to be consistent with ECF we are making them 300 requests/5 minutes.