Update validation for post-separation
Now that we have separated, we want to udpate the `User` and `Application`
validation as we no longer need the `on: :npq_separation` context (the data is
now consistent post-migration).

We also want to relax the `uid` validation on `User` as it is not valid for a
user to change their `uid` to another value.