Fix flakey assurance query spec
The issue seems to be the application schedule is no longer valid for the
application course when we come to update the training status. I haven't been
able to figure out why this is and the `rspec --bisect` returns a _really_ long
combination of specs.

Attempting to fix by updating the declaration `training_status` during creation
in case its time-travel related.