Update validations post separation (#2032)
* Update validation for post-separation

Now that we have separated, we want to udpate the `User` and `Application`
validation as we no longer need the `on: :npq_separation` context (the data is
now consistent post-migration).

We also want to relax the `uid` validation on `User` as it is not valid for a
user to change their `uid` to another value.

* Fix flakey assurance query spec

The issue seems to be the application schedule is no longer valid for the
application course when we come to update the training status. I haven't been
able to figure out why this is and the `rspec --bisect` returns a _really_ long
combination of specs.

Attempting to fix by updating the declaration `training_status` during creation
in case its time-travel related.