Fix schedule spec failures due to date changes (#2098)
* Fix schedule spec failures

After certain dates the schedules change for courses. Create both courses to avoid
failures at certain times in the year.

I avoided the `with_default_schedules` setup as it does not amend dates to stop
declaration failures which is needed in those specs.

* Set default schedules on api doc specs for application

Remove the schedule factory and set all schedules with default
helper to avoid failures. We don't need to set a schedule on accepted
applications as well.

* Ensure spec sets statements with different months

The factory chooses different months for a cohort randomly. Set the months
explicitly to avoid a flakey spec where a statement could be the same month
since a validation was instated to avoid duplicates.