Add workflow for refreshing the Knapsack manifest
As we add/remove tests the manifest becomes gradually outdated and as a result
Knapsack can't distribute the test suite across workers as efficiently.

Add workflow to recompute the Knapsack manifest and raise a pull request with
the changes. This can be triggered manually and will also run on a weekly
schedule.

The action to create a pull request also supports assigning a team, but in
order to do this you need to generate a personal access token. To avoid
coupling the action to a user account I havent done this, but if we have access
to a generic 'dfe user' we could do this in the future.