Add workflow for refreshing the Knapsack manifest (#2124)
* Remove flakey Rake task spec

This test for a Rake task was failing when trying to run all tests sequentially
to build the Knapsack manifest.

Remove Rake task specs and move testable logic into the underlying service.
Update calls to service to use the new interface.

Move test to ensure user being merged is reloaded into relevant service.

* Add workflow for refreshing the Knapsack manifest

As we add/remove tests the manifest becomes gradually outdated and as a result
Knapsack can't distribute the test suite across workers as efficiently.

Add workflow to recompute the Knapsack manifest and raise a pull request with
the changes. This can be triggered manually and will also run on a weekly
schedule.

The action to create a pull request also supports assigning a team, but in
order to do this you need to generate a personal access token. To avoid
coupling the action to a user account I havent done this, but if we have access
to a generic 'dfe user' we could do this in the future.

* Revert "Remove flakey Rake task spec"

This reverts commit c68ba686d35c14a504cbd2a1b44b1d6d0da59632.

* Fix failing Rake specs

The Rake specs are failing when ran sequentially in CI; loading the Rake
tasks a single time before the suite runs appears to resolve the issue.

* Switch to GitHub app auth

This will allow the CI checks to run on the raised PRs.