[CPDNPQ-2466] Replace use of non-deterministic data in factories (#2164)
* [CPDNPQ-2466] Remove random behaviour from factories

This leads to non-deterministic specs, if they prove to be flaky. If
that happens in CI then debugging the issue becomes very difficult
because of parallelisation and inability to replicated the failure.

* [CPDNPQ-2466] fix flaky statement query spec

Underlying the issue was the `:next_period` trait which is no longer
required and is will potentially collide with the sequenced periods.

* [CPDNPQ-2466] Fix 'next period' assumption in spec

* [CPDNPQ-2466] Renamed trait for clarity

Makes clear that trait is not intended for use in specs

* [CPDNPQ-2466] Fixed flaky users admin page spec

Page results are ordered by users email, spec assumes later emails will
be lexically higher but that may not always be the case