[CPDNPQ-2753] Disable nonce-compatibility mode
This is causing noise in the logs when trying to debug real issues. Its
there to support _really_ old browsers and the option to turn it off has
been there for the last 6 years.

All modern browsers have supported nonces for many many versions now so
it should be safe to remove this fallback mode and silence some warnings