[CPDNPQ-2930] Rework k8s liveness probe (#2654)
### Context

Ticket:
[CPDNPQ-2930](https://dfedigital.atlassian.net/browse/CPDNPQ-2930)

During the disaster recovery test it was found the application was
failing to boot with an empty database. This complicated DR recovery.

### Changes proposed in this pull request

1. Remove the review app specific workaround for the failing to boot
with an empty database
2. Change the failure response from the healthcheck to be service
unavailable instead of the generic error
3. Added a more minimal `/up` endpoint which only checks the app has
booted and connected to the DB
4. Added Redis connection checks to the more thorough `/healthcheck`
endpoint

### Failed feature specs screenshots

If any of the feature specs would fail, there will be page on the wiki
with all
failures:

https://github.com/DFE-Digital/npq-registration/wiki/


[CPDNPQ-2930]:
https://dfedigital.atlassian.net/browse/CPDNPQ-2930?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ