[CPDNPQ-2812] handle error scenario (#2662)
### Context

Ticket: https://dfedigital.atlassian.net/browse/CPDNPQ-2812
The main issue is that clicking 'Back' on our steps ends up going to the
wrong page.

To reproduce the error in this ticket:
* go through journey answering 'No' at RTTA step: `course-start-date` ->
`provider-check` -> `teacher-catchment` -> `work-setting` ('Other') ->
`referred-by-return-to-teaching-adviser` ('No') -> `choose-your-npq`
('senior-leadership') -> `ineligible-for-funding` -> `funding-your-npq`
* click 'Back' until you get to the work-setting: `funding-your-npq` ->
`ineligible-for-funding` -> `choose-your-npq` -> `work-setting`
* go forwards again, but answer 'Yes' at the RTTA step:` work-setting`
-> `referred-by-return-to-teaching-adviser` (yes) -> `choose-your-npq
`-> `possible-funding` -> `choose-your-provider`
* click 'Back' - you end up at `funding-your-npq` - which is wrong
* click 'Back' again - you end up at `ineligible-for-funding` - which is
wrong, and raises an error

### Changes proposed in this pull request

The fix in this PR is just a quick fix, for the scenario above - I have
not corrected the steps, but when the `ineligible-for-funding` step
raises an error, the user will be redirected to the `course-start-date`
step.