[CPDNPQ-2811] application accept error messages (#2708)
### Context

Ticket: https://dfedigital.atlassian.net/browse/CPDNPQ-2811

currently, 3 scenarios are covered by the "Schedule cannot be
determined" error:
1. A schedule identifier has been specified when accepting the
application, but that schedule does not exist in the application's
cohort - this is something under the user's control - they can specify a
different identifier, or not specify one, and we'll determine the
correct schedule.
2. No schedule identifier has been specified, and we have not been able
to determine what schedule to use - this is a problem on our end.
3. When the selected schedule is not valid for the application's course,
we currently show 2 errors: "The selected schedule is not valid for the
course" and "Schedule cannot be determined" - we shouldn't be showing
the second error really

### Changes proposed in this pull request

- For scenario (1.) above - I'm suggesting we should have a better
error: "The selected schedule cannot be found in the application cohort"
- For scenario (2.) we still have the same error message, and we still
send an error to Sentry, as this is something we need to look into.
- For scenario (3.) we will only show the one error: "The selected
schedule is not valid for the course"
- Make our seeded schedules look more like production