[CPDNPQ-3136] Fix bin/dev: add foreman to Gemfile (#2764)
### Context

Ticket: https://dfedigital.atlassian.net/browse/CPDNPQ-3136

36fbf66 changed runtime permissions in the container such that `gem
install` doesn't work.

This means that running `bin/dev` in a container will fail if the
`foreman` gem is not already installed.

By adding it to `Gemfile` in the development group, it'll be installed
when the image is built, so there's no need to install it at runtime,
and the permissions issue is avoided.