[CPDNPQ-3028] Update funding eligibility calculator with latest criteria (#2775)
### Context

Ticket: https://dfedigital.atlassian.net/browse/CPDNPQ-3028

Implements 2025 funding eligibility matrix (link to that document is in
the ticket)

Apologies for the big PR... it may be easiest to review using the
"Commits" dropdown in the "Files changed" tab, I've tried to keep the
commits such that they each make sense individually!

It's a bit oversized because there's a fair bit of refactoring in
here... I found it quite taxing to figure out what was going on with the
existing algorithm, let alone update it with confidence. So given we're
now fully in charge of our product, and this year's eligibility criteria
is a significant divergence from last year's, it seemed like an
appropriate moment to align the eligibility algorithm with the way the
policy is illustrated in the matrix:

- The paths in `FundingEligibility` are now determined by work setting
and then any other criteria, this aligns it with the matrix document
_and_ the user journey through the wizard (so hopefully easier to
understand/troubleshoot/amend), and also helps guard us from
stale/incoherent answers that can be in the store if people change
answers in the journey
- The spec is completely rewritten to also be aligned with the
eligibility matrix's criteria: each possible funding scenario now has
its own context that explicitly specifies the expected status for every
course, so again easier to follow against the matrix / user journey
- Funding decisions that are subject to review are now treated as an
explicit status code, rather than being inferred elsewhere from various
answers
- Steps whose `next_step` depends on funding policy now refer to the
`FundingEligibility` class too, rather than having their own separate
implementation of the policy scattered throughout the steps themselves,
this should make the wizard path and application funding status
coherent, plus make future updates easier
- Specs no longer use `stub_const` to mock institution
inclusion/exclusion on lists, instead they stub the predicate methods on
the institution instance, so that the specs are decoupled from the way
eligibility lists happen to be implemented right now and independent of
whether any factory-generated URNs happen to be on any lists.

There's definitely more that we could do to improve this further, but
I've brought down the scope hammer here so that we can start testing
imminently ready for re-opening.