[CPDNPQ-3028] Follow-up 1: fix childminder eligibility (#2819)
### Context

Ticket: https://dfedigital.atlassian.net/browse/CPDNPQ-3028

8f104b0e7f51ae2852e1996ee461ef2dca68235f: Testing has revealed a
misunderstanding in the criteria for childminders to qualify for the EYL
NPQ. In previous cohorts, a childminder's URN had to appear on a special
eligibility list to qualify. This is no longer the case, there is no
such eligibility list for this cohort and we are marking all
Ofsted-registered childminders as eligible for funding.

2a41b209bb3c996510d298c02597444712e2834f: As a result, the method and
initializer for this list is redundant so we can also remove it to avoid
ambiguity going forwards. I did consider keeping it just in case the
list comes back in the future, but it's trivial to reimplement and we
have a new approach for such lists anyway, so it seems worth bringing
down the axe.